FEAT(*): 친밀도 및 관계 디테일 뷰
1. 관계 디테일 API 호출 시 어떤 행동들이 엮여있는지 나오도록 추가
2. 일기 디테일 API 호출 시 친밀도 몇 점이 감산/가산되었는지 나오도록 추가

+++ src/entities/diary-target.entity.ts
@@ -21,7 +21,7 @@ export class DiaryTarget {
   @JoinColumn({ name: 'target_id' })
   target!: Target;
   
-  @Column({ type: 'int', default: 0 })
+  @Column({ type: 'float', default: 0 })
   changeScore:number
